pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "vewutils"
authors = [
    {name = "Shintaro Bunya", email = "sbunya@gmail.com"}
]
description = "Utility programs for the ADCIRC hydrodynamic model"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pandas",
    "numpy",
    "shapely",
    "plotly",
    "geopandas",
    "scipy",
    "rasterio",
    "rasterstats",
    "geowombat",
    "pyproj>=2.6",
    "pyyaml",
    "matplotlib",
    "xarray",
    "requests",
    "pytz",
    "jupyter",
    "ipywidgets",
    "netCDF4",
    "haversine",
    "paramiko",
    "pooch",
    "psutil",
    "searvey",
    "typepigeon<2",
    "utm",
    "appdirs"
]

[project.optional-dependencies]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
    "myst-parser"
]
